+++ PythonAPI/carla/agents/navigation/global_route_planner.py
@@ -31,6 +31,8 @@ class GlobalRoutePlanner(object):
         self._graph = None
         self._id_map = None
         self._road_id_to_edge = None
+        self._intersection_end_node = -1
+        self._previous_decision = RoadOption.VOID
 
     def setup(self):
         """
@@ -39,6 +41,7 @@ class GlobalRoutePlanner(object):
         """
         self._topology = self._dao.get_topology()
         self._graph, self._id_map, self._road_id_to_edge = self._build_graph()
+        self._find_loose_ends()
         self._lane_change_link()
 
     def _build_graph(self):
@@ -67,7 +70,7 @@ class GlobalRoutePlanner(object):
             path = segment['path']
             entry_wp, exit_wp = segment['entry'], segment['exit']
             intersection = entry_wp.is_junction
-            road_id, lane_id = entry_wp.road_id, entry_wp.lane_id
+            road_id, section_id, lane_id = entry_wp.road_id, entry_wp.section_id, entry_wp.lane_id
 
             for vertex in entry_xyz, exit_xyz:
                 # Adding unique nodes and populating id_map
@@ -79,23 +82,69 @@ class GlobalRoutePlanner(object):
             n2 = id_map[exit_xyz]
             if road_id not in road_id_to_edge:
                 road_id_to_edge[road_id] = dict()
-            road_id_to_edge[road_id][lane_id] = (n1, n2)
+            if section_id not in road_id_to_edge[road_id]:
+                road_id_to_edge[road_id][section_id] = dict()
+            road_id_to_edge[road_id][section_id][lane_id] = (n1, n2)
 
+            entry_carla_vector = entry_wp.transform.rotation.get_forward_vector()
+            exit_carla_vector = exit_wp.transform.rotation.get_forward_vector()
             # Adding edge with attributes
             graph.add_edge(
                 n1, n2,
                 length=len(path) + 1, path=path,
                 entry_waypoint=entry_wp, exit_waypoint=exit_wp,
-                entry_vector=vector(
-                    entry_wp.transform.location,
-                    path[0].transform.location if len(path) > 0 else exit_wp.transform.location),
-                exit_vector=vector(
-                    path[-1].transform.location if len(path) > 0 else entry_wp.transform.location,
-                    exit_wp.transform.location),
+                entry_vector=np.array(
+                    [entry_carla_vector.x, entry_carla_vector.y, entry_carla_vector.z]),
+                exit_vector=np.array(
+                    [exit_carla_vector.x, exit_carla_vector.y, exit_carla_vector.z]),
                 net_vector=vector(entry_wp.transform.location, exit_wp.transform.location),
                 intersection=intersection, type=RoadOption.LANEFOLLOW)
 
         return graph, id_map, road_id_to_edge
+
+    def _find_loose_ends(self):
+        """
+        This method finds road segments that have an unconnected end and
+        adds them to the internal graph representation
+        """
+        count_loose_ends = 0
+        hop_resolution = self._dao.get_resolution()
+        for segment in self._topology:
+            end_wp = segment['exit']
+            exit_xyz = segment['exitxyz']
+            road_id, section_id, lane_id = end_wp.road_id, end_wp.section_id, end_wp.lane_id
+            if road_id in self._road_id_to_edge and \
+                section_id in self._road_id_to_edge[road_id] and \
+                    lane_id in self._road_id_to_edge[road_id][section_id]:
+                pass
+            else:
+                count_loose_ends += 1
+                if road_id not in self._road_id_to_edge:
+                    self._road_id_to_edge[road_id] = dict()
+                if section_id not in self._road_id_to_edge[road_id]:
+                    self._road_id_to_edge[road_id][section_id] = dict()
+                n1 = self._id_map[exit_xyz]
+                n2 = -1*count_loose_ends
+                self._road_id_to_edge[road_id][section_id][lane_id] = (n1, n2)
+                next_wp = end_wp.next(hop_resolution)
+                path = []
+                while next_wp is not None and next_wp and \
+                    next_wp[0].road_id == road_id and \
+                        next_wp[0].section_id == section_id and \
+                            next_wp[0].lane_id == lane_id:
+                    path.append(next_wp[0])
+                    next_wp = next_wp[0].next(hop_resolution)
+                if path:
+                    n2_xyz = (path[-1].transform.location.x,
+                              path[-1].transform.location.y,
+                              path[-1].transform.location.z)
+                    self._graph.add_node(n2, vertex=n2_xyz)
+                    self._graph.add_edge(
+                        n1, n2,
+                        length=len(path) + 1, path=path,
+                        entry_waypoint=end_wp, exit_waypoint=path[-1],
+                        entry_vector=None, exit_vector=None, net_vector=None,
+                        intersection=end_wp.is_intersection, type=RoadOption.LANEFOLLOW)
 
     def _localize(self, location):
         """
@@ -104,7 +153,18 @@ class GlobalRoutePlanner(object):
         return          :   pair node ids representing an edge in the graph
         """
         waypoint = self._dao.get_waypoint(location)
-        return self._road_id_to_edge[waypoint.road_id][waypoint.lane_id]
+        edge = None
+        try:
+            edge = self._road_id_to_edge[waypoint.road_id][waypoint.section_id][waypoint.lane_id]
+        except KeyError:
+            print(
+                "Failed to localize! : ",
+                "Road id : ", waypoint.road_id,
+                "Section id : ", waypoint.section_id,
+                "Lane id : ", waypoint.lane_id,
+                "Location : ", waypoint.transform.location.x,
+                waypoint.transform.location.y)
+        return edge
 
     def _lane_change_link(self):
         """
@@ -121,13 +181,11 @@ class GlobalRoutePlanner(object):
 
                     if bool(waypoint.lane_change & carla.LaneChange.Right) and not right_found:
                         next_waypoint = waypoint.get_right_lane()
-                        if next_waypoint is not None and next_waypoint.lane_type == carla.LaneType.Driving and \
-                            waypoint.road_id == next_waypoint.road_id:
+                        if next_waypoint is not None and \
+                            next_waypoint.lane_type == carla.LaneType.Driving and \
+                                waypoint.road_id == next_waypoint.road_id:
                             next_road_option = RoadOption.CHANGELANERIGHT
-                            try:
-                                next_segment = self._localize(next_waypoint.transform.location)
-                            except KeyError:
-                                print("Failed to localize! : ", next_waypoint.road_id, next_waypoint.lane_id)
+                            next_segment = self._localize(next_waypoint.transform.location)
                             if next_segment is not None:
                                 self._graph.add_edge(
                                     self._id_map[segment['entryxyz']], next_segment[0], entry_waypoint=segment['entry'],
@@ -140,10 +198,7 @@ class GlobalRoutePlanner(object):
                         if next_waypoint is not None and next_waypoint.lane_type == carla.LaneType.Driving and \
                             waypoint.road_id == next_waypoint.road_id:
                             next_road_option = RoadOption.CHANGELANELEFT
-                            try:
-                                next_segment = self._localize(next_waypoint.transform.location)
-                            except KeyError:
-                                print("Failed to localize! : ", next_waypoint.road_id, next_waypoint.lane_id)
+                            next_segment = self._localize(next_waypoint.transform.location)
                             if next_segment is not None:
                                 self._graph.add_edge(
                                     self._id_map[segment['entryxyz']], next_segment[0], entry_waypoint=segment['entry'],
@@ -181,6 +236,30 @@ class GlobalRoutePlanner(object):
         route.append(end[1])
         return route
 
+    def _successive_last_intersection_edge(self, index, route):
+        """
+        This method returns the last successive intersection edge
+        from a starting index on the route.
+
+        This helps moving past tiny intersection edges to calculate
+        proper turn decisions.
+        """
+
+        last_intersection_edge = None
+        last_node = None
+        for node1, node2 in [(route[i], route[i+1]) for i in range(index, len(route)-1)]:
+            candidate_edge = self._graph.edges[node1, node2]
+            if node1 == route[index]:
+                last_intersection_edge = candidate_edge
+            if candidate_edge['type'] == RoadOption.LANEFOLLOW and \
+                candidate_edge['intersection']:
+                last_intersection_edge = candidate_edge
+                last_node = node2
+            else:
+                break
+
+        return last_node, last_intersection_edge
+
     def _turn_decision(self, index, route, threshold=math.radians(5)):
         """
         This method returns the turn decision (RoadOption) for pair of edges
@@ -193,35 +272,52 @@ class GlobalRoutePlanner(object):
         next_node = route[index+1]
         next_edge = self._graph.edges[current_node, next_node]
         if index > 0:
-            current_edge = self._graph.edges[previous_node, current_node]
-            calculate_turn = current_edge['type'].value == RoadOption.LANEFOLLOW.value and \
-                not current_edge['intersection'] and \
-                    next_edge['type'].value == RoadOption.LANEFOLLOW.value and \
-                        next_edge['intersection']
-            if calculate_turn:
-                cv, nv = current_edge['exit_vector'], next_edge['net_vector']
-                cross_list = []
-                for neighbor in self._graph.successors(current_node):
-                    select_edge = self._graph.edges[current_node, neighbor]
-                    if select_edge['type'].value == RoadOption.LANEFOLLOW.value:
-                        if neighbor != route[index+1]:
-                            sv = select_edge['net_vector']
-                            cross_list.append(np.cross(cv, sv)[2])
-                next_cross = np.cross(cv, nv)[2]
-                deviation = math.acos(np.dot(cv, nv) /\
-                    (np.linalg.norm(cv)*np.linalg.norm(nv)))
-                if not cross_list:
-                    cross_list.append(0)
-                if deviation < threshold:
-                    decision = RoadOption.STRAIGHT
-                elif cross_list and next_cross < min(cross_list):
-                    decision = RoadOption.LEFT
-                elif cross_list and next_cross > max(cross_list):
-                    decision = RoadOption.RIGHT
+            if self._previous_decision != RoadOption.VOID and \
+                self._intersection_end_node > 0 and \
+                    self._intersection_end_node != previous_node and \
+                        next_edge['type'] == RoadOption.LANEFOLLOW and \
+                            next_edge['intersection']:
+                decision = self._previous_decision
             else:
-                decision = next_edge['type']
+                self._intersection_end_node = -1
+                current_edge = self._graph.edges[previous_node, current_node]
+                calculate_turn = current_edge['type'].value == RoadOption.LANEFOLLOW.value and \
+                    not current_edge['intersection'] and \
+                        next_edge['type'].value == RoadOption.LANEFOLLOW.value and \
+                            next_edge['intersection']
+                if calculate_turn:
+                    last_node, tail_edge = self._successive_last_intersection_edge(index, route)
+                    self._intersection_end_node = last_node
+                    if tail_edge is not None:
+                        next_edge = tail_edge
+                    cv, nv = current_edge['exit_vector'], next_edge['net_vector']
+                    cross_list = []
+                    for neighbor in self._graph.successors(current_node):
+                        select_edge = self._graph.edges[current_node, neighbor]
+                        if select_edge['type'].value == RoadOption.LANEFOLLOW.value:
+                            if neighbor != route[index+1]:
+                                sv = select_edge['net_vector']
+                                cross_list.append(np.cross(cv, sv)[2])
+                    next_cross = np.cross(cv, nv)[2]
+                    deviation = math.acos(np.clip(
+                        np.dot(cv, nv)/(np.linalg.norm(cv)*np.linalg.norm(nv)), -1.0, 1.0))
+                    if not cross_list:
+                        cross_list.append(0)
+                    if deviation < threshold:
+                        decision = RoadOption.STRAIGHT
+                    elif cross_list and next_cross < min(cross_list):
+                        decision = RoadOption.LEFT
+                    elif cross_list and next_cross > max(cross_list):
+                        decision = RoadOption.RIGHT
+                    elif next_cross < 0:
+                        decision = RoadOption.LEFT
+                    elif next_cross > 0:
+                        decision = RoadOption.RIGHT
+                else:
+                    decision = next_edge['type']
         else:
             decision = next_edge['type']
+        self._previous_decision = decision
 
         return decision
 
@@ -259,12 +355,14 @@ class GlobalRoutePlanner(object):
 
     def trace_route(self, origin, destination):
         """
-        This method returns list of (carla.Waypoint, RoadOption) from origin to destination
+        This method returns list of (carla.Waypoint, RoadOption)
+        from origin (carla.Location) to destination (carla.Location)
         """
 
         route_trace = []
         route = self._path_search(origin, destination)
         current_waypoint = self._dao.get_waypoint(origin)
+        destination_waypoint = self._dao.get_waypoint(destination)
         resolution = self._dao.get_resolution()
 
         for i in range(len(route) - 1):
@@ -272,8 +370,11 @@ class GlobalRoutePlanner(object):
             edge = self._graph.edges[route[i], route[i+1]]
             path = []
 
-            if edge['type'].value != RoadOption.LANEFOLLOW.value and edge['type'].value != RoadOption.VOID.value:
-                n1, n2 = self._road_id_to_edge[edge['exit_waypoint'].road_id][edge['exit_waypoint'].lane_id]
+            if edge['type'].value != RoadOption.LANEFOLLOW.value and \
+                edge['type'].value != RoadOption.VOID.value:
+                route_trace.append((current_waypoint, road_option))
+                exit_wp = edge['exit_waypoint']
+                n1, n2 = self._road_id_to_edge[exit_wp.road_id][exit_wp.section_id][exit_wp.lane_id]
                 next_edge = self._graph.edges[n1, n2]
                 if next_edge['path']:
                     closest_index = self._find_closest_in_list(current_waypoint, next_edge['path'])
@@ -289,7 +390,15 @@ class GlobalRoutePlanner(object):
                 for waypoint in path[closest_index:]:
                     current_waypoint = waypoint
                     route_trace.append((current_waypoint, road_option))
-                    if len(route)-i <= 2 and waypoint.transform.location.distance(destination) < 2*resolution:
+                    if len(route)-i <= 2 and \
+                        waypoint.transform.location.distance(destination) < 2*resolution:
                         break
+                    elif len(route)-i <= 2 and \
+                        current_waypoint.road_id == destination_waypoint.road_id and \
+                            current_waypoint.section_id == destination_waypoint.section_id and \
+                                current_waypoint.lane_id == destination_waypoint.lane_id:
+                        destination_index = self._find_closest_in_list(destination_waypoint, path)
+                        if closest_index > destination_index:
+                            break
 
         return route_trace