+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Recorder/CarlaRecorderHelpers.h
@@ -8,6 +8,9 @@
 
 #include <fstream>
 
+// get the final path + filename
+std::string GetRecorderFilename(std::string Filename);
+
 // ---------
 // recorder
 // ---------