+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Recorder/CarlaRecorderQuery.h
@@ -7,12 +7,14 @@
 #pragma once
 
 #include <fstream>
-#include "CarlaRecorderInfo.h"
-#include "CarlaRecorderFrames.h"
+
+#include "CarlaRecorderAnimWalker.h"
+#include "CarlaRecorderCollision.h"
 #include "CarlaRecorderEventAdd.h"
 #include "CarlaRecorderEventDel.h"
 #include "CarlaRecorderEventParent.h"
-#include "CarlaRecorderCollision.h"
+#include "CarlaRecorderFrames.h"
+#include "CarlaRecorderInfo.h"
 #include "CarlaRecorderPosition.h"
 #include "CarlaRecorderState.h"
 
@@ -48,6 +50,8 @@ private:
   CarlaRecorderPosition Position;
   CarlaRecorderCollision Collision;
   CarlaRecorderStateTrafficLight StateTraffic;
+  CarlaRecorderAnimVehicle Vehicle;
+  CarlaRecorderAnimWalker Walker;
 
   // read next header packet
   bool ReadHeader(void);