+++ CHANGELOG.md
@@ -1,7 +1,21 @@
-## Latest Changes
+## Latest
 
+  * Recorder fixes:
+    - Actors at start of playback could interpolate positions from its current position instead than the recorded position, making some fast sliding effect during 1 frame.
+    - Camera following in playback was not working if a new map was needed to load.
+    - API function 'show_recorder_file_info' was showing the wrong parent id.
+    - Script 'start_recording.py' now properly saves destruction of actors at stop.
   * API extension: waypoint's `junction_id` that returns de OpenDrive identifier of the current junction
   * API change: deprecated waypoint's `is_intersection`, now is `is_junction`
+  * New recorder features:
+    - Added optional parameter to show more details about a recorder file (related to `show_recorder_file_info.py`)
+    - Added playback speed (slow/fast motion) for the replayer
+    - We can use an absolute path for the recorded files (to choose where to 'write to' or 'read from')
+    - New data recorded to replay animations:
+      - Wheels of vehicles are animated (steering, throttle, handbrake), also bikes and cycles
+      - Walkers animation is simulated in playback (through speed of walker), so they walk properly.
+  * Fixed Lidar effectiveness bug in manual_control.py
+  * Added C++ client example using LibCarla
 
 ## CARLA 0.9.5